requirements.txt
allure_python_commons==2.15.0
jsonpath==0.82.2
pymysql==1.1.2
pytest==8.4.2
PyYAML==6.0.3
Requests==2.32.5
tqdm==4.67.1
